[Core] Renamed exception.hpp to errors.hpp.
Since there's only one exception, but a lot of error codes.
