+++ include/cocaine/errors.hpp
