+++ include/cocaine/rpc/asio/readable_stream.hpp
@@ -21,7 +21,7 @@
 #ifndef COCAINE_IO_BUFFERED_READABLE_STREAM_HPP
 #define COCAINE_IO_BUFFERED_READABLE_STREAM_HPP
 
-#include "cocaine/exceptions.hpp"
+#include "cocaine/errors.hpp"
 
 #include <functional>
